add stats and coorect lots of bugs

+++ transcendence/website/views.py
@@ -9,13 +9,18 @@ from django.shortcuts import get_object_or_404
 from django.http import HttpResponseRedirect
 from django.urls import reverse
 from django.contrib.auth.forms import AuthenticationForm
+from django.http import JsonResponse
+from django.contrib.auth.decorators import login_required
 
 # Create your views here.
 def main(request):
     password_changed = 'password_changed' in request.GET
+    avatar_changed = 'avatar_changed' in request.GET
+    update_session_auth_hash(request, request.user)
     context = {
         'user': request.user,
-        'password_changed': password_changed,  # Ajout de cette ligne
+        'password_changed': password_changed,
+        'avatar_changed': avatar_changed,
     }
     return render(request, 'main.html', context)
 
@@ -89,12 +94,12 @@ def logout_view(request):
 from django.contrib import messages
 
 def profile(request):
+    user_stats = User.objects.get(id=request.user.id)
     if request.method == 'POST':
         pwd_form = ChangePasswordForm(request.user, request.POST)
-        avatar_form = ChangeAvatarForm(request.user, request.POST, request.FILES)
+        avatar_form = ChangeAvatarForm(request.POST, request.FILES)
 
         if 'change_password' in request.POST:
-            # Traitement du formulaire de changement de mot de passe
             if pwd_form.is_valid():
                 pwd_form.save()
                 update_session_auth_hash(request, request.user)
@@ -102,27 +107,46 @@ def profile(request):
                 return redirect('/main?password_changed=true')
             else:
                 messages.error(request, 'Error in password change form submission')
+        
         elif 'change_avatar' in request.POST:
-            old_avatar = request.user.avatar
-            if old_avatar and not old_avatar.name.endswith('default.png'):
-                old_avatar.delete()
-            request.user.avatar = request.FILES['avatar']
+            if avatar_form.is_valid():
+                old_avatar = request.user.avatar
+                if old_avatar and not old_avatar.name.endswith('default.png'):
+                    old_avatar.delete()
+                request.user.avatar = request.FILES['avatar']
+                request.user.save()
+                update_session_auth_hash(request, request.user)
+                messages.success(request, 'Avatar changed successfully')
+                return redirect('/main?avatar_changed=true')
+            else:
+                messages.error(request, 'Error in avatar change form submission')
+
+        elif 'change_name' in request.POST:
+            # Aucun formulaire n'est nécessaire car vous récupérez directement les données du POST
+            request.user.first_name = request.POST['first_name']
+            request.user.last_name = request.POST['last_name']
             request.user.save()
             update_session_auth_hash(request, request.user)
-            messages.success(request, 'Avatar changed successfully')
-            return redirect('/main?avatar_changed=true')
+            messages.success(request, 'Name changed successfully')
+            return redirect('/main')
         else:
-            messages.error(request, 'Error in avatar change form submission')
-            print(avatar_form.errors)
+            messages.error(request, 'An unexpected error occurred')
+
     else:
         pwd_form = ChangePasswordForm(request.user)
-        avatar_form = ChangeAvatarForm(request.user)
+        avatar_form = ChangeAvatarForm()
 
-    return render(request, 'profile.html', {'pwd_form': pwd_form, 'avatar_form': avatar_form})
+    return render(request, 'profile.html', {'pwd_form': pwd_form, 'avatar_form': avatar_form, 'user_stats': user_stats})
 
 
 def handler404(request, exception):
     return render(request, '404.html', status=404)
+
+def increment_game(request, player_id):
+    player = get_object_or_404(User, id=player_id)
+    player.total_pong_games += 1
+    player.save()
+    return HttpResponseRedirect(reverse('details', args=[player_id]))
 
 def increment_victory(request, player_id):
     player = get_object_or_404(User, id=player_id)
@@ -160,3 +184,7 @@ def resetWL(request, player_id):
 def pong(request):
     # Your view logic here
     return render(request, 'pong.html')
+
+@login_required
+def	check_login(request):
+    return JsonResponse({'is_logged_in': request.user.is_authenticated})
